+++ affirm/src/main/java/com/affirm/android/AffirmConstants.java
@@ -15,7 +15,7 @@ public final class AffirmConstants {
     static final String CHECKOUT_PATH = "/api/v2/checkout/";
     static final String TRACKER_PATH = "/collect";
     static final String PROMO_PATH = "/api/promos/v2/%s?is_sdk=true&field=ala&amount=%d"
-            + "&show_cta=%s&promo_external_id=%s";
+            + "&show_cta=%s";
     static final String PREQUAL_PATH = "/apps/prequal?public_api_key=%s&unit_price=%s"
             + "&promo_external_id=%s&isSDK=true&use_promo=True&referring_url=%s";
     static final String AFFIRM_CHECKOUT_CONFIRMATION_URL = "affirm://checkout/confirmed";